requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-allauth==0.58.2
django-tinymce==3.6.1
django-betterforms==2.0.0